rename internal repository to StatementStorage
